Last bits of cleanup with pytest configuration for E2E tests
